init maven project + directory organization
