+++ README.md
@@ -1,1 +1,4 @@
 # tup2024
+
+execute with `java -cp target/sistema-bancario-1.0-SNAPSHOT.jar ar.edu.utn.frbb.tup.Aplicacion`
+
