--- src/ar/edu/utn/frbb/tup/Aplicacion.java
@@ -1,25 +1,0 @@
-package ar.edu.utn.frbb.tup;
-
-import ar.edu.utn.frbb.tup.utils.Cliente;
-import ar.edu.utn.frbb.tup.utils.Cuenta;
-import ar.edu.utn.frbb.tup.utils.SimpleAdder;
-import ar.edu.utn.frbb.tup.utils.TipoPersona;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Scanner;
-
-class Aplicacion {
-
-    public static void main(String args[]) {
-        Banco banco = new Banco();
-
-        MenuInputProcessor menuInputProcessor = new MenuInputProcessor();
-        menuInputProcessor.renderMenu(banco);
-
-    }
-
-
-}
