--- src/ar/edu/utn/frbb/tup/Banco.java
@@ -1,18 +1,0 @@
-package ar.edu.utn.frbb.tup;
-
-import ar.edu.utn.frbb.tup.utils.Cliente;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Banco {
-    private List<Cliente> clientes = new ArrayList<>();
-
-    public List<Cliente> getClientes() {
-        return clientes;
-    }
-
-    public void setClientes(List<Cliente> clientes) {
-        this.clientes = clientes;
-    }
-}
