--- src/ar/edu/utn/frbb/tup/BaseInputProcessor.java
@@ -1,14 +1,0 @@
-package ar.edu.utn.frbb.tup;
-
-import java.util.Scanner;
-
-public class BaseInputProcessor {
-
-    protected Scanner scanner = new Scanner(System.in);
-
-
-    protected static void clearScreen() {
-        System.out.print("\033[H\033[2J");
-        System.out.flush();
-    }
-}
