+++ src/main/java/ar/edu/utn/frbb/tup/model/Banco.java
@@ -1,0 +1,16 @@
+package ar.edu.utn.frbb.tup.model;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Banco {
+  private List<Cliente> clientes = new ArrayList<>();
+
+  public List<Cliente> getClientes() {
+    return clientes;
+  }
+
+  public void setClientes(List<Cliente> clientes) {
+    this.clientes = clientes;
+  }
+}
