+++ src/main/java/ar/edu/utn/frbb/tup/model/Cuenta.java
@@ -1,0 +1,36 @@
+package ar.edu.utn.frbb.tup.model;
+
+import java.time.LocalDateTime;
+
+public class Cuenta {
+  String nombre;
+  LocalDateTime fechaCreacion;
+  int balance;
+
+  public String getNombre() {
+    return nombre;
+  }
+
+  public Cuenta setNombre(String nombre) {
+    this.nombre = nombre;
+    return this;
+  }
+
+  public LocalDateTime getFechaCreacion() {
+    return fechaCreacion;
+  }
+
+  public Cuenta setFechaCreacion(LocalDateTime fechaCreacion) {
+    this.fechaCreacion = fechaCreacion;
+    return this;
+  }
+
+  public int getBalance() {
+    return balance;
+  }
+
+  public Cuenta setBalance(int balance) {
+    this.balance = balance;
+    return this;
+  }
+}
