+++ src/main/java/ar/edu/utn/frbb/tup/model/Persona.java
@@ -1,0 +1,42 @@
+package ar.edu.utn.frbb.tup.model;
+
+import java.time.LocalDate;
+
+public class Persona {
+  private String nombre;
+  private String apellido;
+  private long dni;
+  private LocalDate fechaNacimiento;
+
+  public String getNombre() {
+    return nombre;
+  }
+
+  public void setNombre(String nombre) {
+    this.nombre = nombre;
+  }
+
+  public String getApellido() {
+    return apellido;
+  }
+
+  public void setApellido(String apellido) {
+    this.apellido = apellido;
+  }
+
+  public long getDni() {
+    return dni;
+  }
+
+  public void setDni(long dni) {
+    this.dni = dni;
+  }
+
+  public LocalDate getFechaNacimiento() {
+    return fechaNacimiento;
+  }
+
+  public void setFechaNacimiento(LocalDate fechaNacimiento) {
+    this.fechaNacimiento = fechaNacimiento;
+  }
+}
