+++ src/main/java/ar/edu/utn/frbb/tup/model/TipoPersona.java
@@ -1,0 +1,26 @@
+package ar.edu.utn.frbb.tup.model;
+
+public enum TipoPersona {
+  PERSONA_FISICA("F"),
+  PERSONA_JURIDICA("J");
+
+  private final String descripcion;
+
+  TipoPersona(String descripcion) {
+    this.descripcion = descripcion;
+  }
+
+  public String getDescripcion() {
+    return descripcion;
+  }
+
+  public static TipoPersona fromString(String text) {
+    for (TipoPersona tipo : TipoPersona.values()) {
+      if (tipo.descripcion.equalsIgnoreCase(text)) {
+        return tipo;
+      }
+    }
+    throw new IllegalArgumentException(
+        "No se pudo encontrar un TipoPersona con la descripción: " + text);
+  }
+}
