+++ src/main/java/ar/edu/utn/frbb/tup/view/ClienteInputProcessor.java
@@ -1,0 +1,55 @@
+package ar.edu.utn.frbb.tup.view;
+
+import ar.edu.utn.frbb.tup.model.Cliente;
+import ar.edu.utn.frbb.tup.model.TipoPersona;
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
+
+public class ClienteInputProcessor extends BaseInputProcessor {
+
+  private static List<Cliente> clientes = new ArrayList<>();
+
+  public Cliente ingresarCliente() {
+
+    // Ingreso de datos del Cliente
+    Cliente cliente = new Cliente();
+    clearScreen();
+    System.out.println("Ingrese el nombre del cliente:");
+    String nombre = scanner.nextLine();
+    cliente.setNombre(nombre);
+
+    System.out.println("Ingrese el apellido del cliente:");
+    String apellido = scanner.nextLine();
+    cliente.setApellido(apellido);
+
+    System.out.println("Ingrese el tipo de persona Física(F) o Jurídica(J):");
+    String tipoPersonaStr = scanner.nextLine().toUpperCase();
+    while (!tipoPersonaStr.equals("F") && !tipoPersonaStr.equals("J")) {
+      System.out.println("Tipo de persona inválido. Ingrese NATURAL o JURIDICA:");
+      tipoPersonaStr = scanner.nextLine().toUpperCase();
+    }
+    TipoPersona tipoPersona = TipoPersona.fromString(tipoPersonaStr);
+    cliente.setTipoPersona(tipoPersona);
+
+    System.out.println("Ingrese el banco del cliente:");
+    String banco = scanner.nextLine();
+    cliente.setBanco(banco);
+
+    System.out.println("Ingrese la fecha de alta del cliente (Formato: YYYY-MM-DD):");
+    LocalDate fechaAlta = null;
+    boolean fechaValida = false;
+    while (!fechaValida) {
+      try {
+        fechaAlta = LocalDate.parse(scanner.nextLine());
+        fechaValida = true;
+      } catch (Exception e) {
+        System.out.println("Formato de fecha inválido. Ingrese la fecha en formato YYYY-MM-DD:");
+      }
+    }
+    cliente.setFechaAlta(fechaAlta);
+
+    clearScreen();
+    return cliente;
+  }
+}
